auto-playwright-tests/requirements.txt
openai>=1.0.0
python-dotenv>=0.19.0
tenacity>=8.0.0
pyyaml>=6.0.1
flask>=2.0.0  # For serving the web form 
requests